\name{ggpca.check.prcomp}
\alias{ggpca.check.prcomp}
\title{Check variance PCA plot (look for specific name for this function)}
\usage{
  ggpca.check.prcomp(pca.fit)
}
\arguments{
  \item{pca.fit}{object of class \code{prcomp}.}
}
\value{
  A ggplot2 object.
}
\description{
  Check variance PCA plot (look for specific name for this
  function)
}
\seealso{
  \code{\link{ggpca.biplot}}
}

