\name{see.na}
\alias{see.na}
\title{Vizualize NAs in a data frame}
\usage{
  see.na(data, num.vars = seq(data), ...)
}
\arguments{
  \item{data}{a data frame.}

  \item{num.vars}{Index (or names) of the numeric variables
  of \code{data}. By default all variables are considered.}

  \item{...}{Additional arguments to \code{\link{see}}}
}
\description{
  Vizualize NAs in a data frame
}

